Create Find the number of islands(six oct POTD -GFG)
This C++ code defines a class Solution that implements a depth-first search (DFS) algorithm to count the number of islands in a 2D grid, where '1' represents land and '0' represents water. The numIslands method iterates through the grid, incrementing a counter each time it finds a '1', and uses the DFS function to mark all connected land cells as '0'.



Kindly assign me this.