Create Count the nodes at distance K from leaf(6 feb GFG-POTD)
This C++ class Solution contains methods to find and count nodes in a binary tree that are at a given distance ( k ) from any leaf node.

solve Method: This recursive function traverses the tree and stores paths from the root to each leaf node in the paths vector.
printKDistantfromLeaf Method: This function uses the paths collected by solve to find nodes that are ( k ) levels above any leaf node. It stores these nodes in an unordered set to ensure uniqueness and returns the size of this set as the count of such nodes.


Please assign me this.